+++ lib/paper_trail/frameworks/rspec/helpers.rb
@@ -4,12 +4,15 @@ module PaperTrail
       # Included in the RSpec configuration in `frameworks/rspec.rb`
       module InstanceMethods
         # enable versioning for specific blocks (at instance-level)
-        def with_versioning
+        def with_versioning(whodunnit: nil)
+          was_whodunnit = ::PaperTrail.whodunnit
           was_enabled = ::PaperTrail.enabled?
+          ::PaperTrail.whodunnit = whodunnit unless whodunnit.nil?
           ::PaperTrail.enabled = true
           yield
         ensure
           ::PaperTrail.enabled = was_enabled
+          ::PaperTrail.whodunnit = was_whodunnit unless whodunnit.nil?
         end
       end